requirements.txt
pymel
scipy
libigl